improve foreach vm loops
@slandelle is there a more idiomatic solution for this?
You can see a ''vapp'' as a collection and ''vm'' as elements in the collection.
I want to add many elements to the collection and then have a way to iterate them.

''saveCreatedVm'' appends to a List session attribute (not sure check#transform can manipulate the session)
''foreachVm'' use repeat to update a ''currentElement'' session attribute

WDYT?